+++ src/lib.rs
@@ -9,10 +9,12 @@ mod sqrt;
 mod vec2;
 mod vec2_range_iterator;
 mod zero;
+mod scale;
 
 pub use self::{
     cast::*, containing_float::*, float::*, goniometric::*, into_float::*,
     isqrt::*, one::*, sqrt::*, vec2::*, vec2_range_iterator::*, zero::*,
+    scale::*,
 };
 
 #[cfg(test)]